+++ lib/screens/common/privacy.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+class PrivacyScreen extends StatefulWidget {
+  const PrivacyScreen({super.key});
+
+  @override
+  State<PrivacyScreen> createState() => _PrivacyScreenState();
+}
+
+class _PrivacyScreenState extends State<PrivacyScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Privacy Policy'),
+      ),
+      body: const Padding(
+        padding: EdgeInsets.all(24.0),
+        child: SingleChildScrollView(
+          child: Text(
+            'Here you can display your privacy policy text.\n\nAdd your content here.',
+            style: TextStyle(fontSize: 16),
+          ),
+        ),
+      ),
+    );
+  }
+}